==> notebooks/13_scooby/1_get_train_data.ipynb <==
# Get the pesudobulk data from scooby

import os
import numpy as np
import pandas as pd
import scipy
import torch
from tqdm import tqdm
import polars as pl
from torch.utils.data import DataLoader
from enformer_pytorch.data import GenomeIntervalDataset
from scooby.data import onTheFlyPseudobulkDataset
from scooby.utils.utils import undo_squashed_scale, get_gene_slice_and_strand
from scooby.utils.transcriptome import Transcriptome

data_path = '/gstore/data/resbioai/grelu/decima/scooby/'  

## Read cell types used for training

cell_type_index = pd.read_parquet(os.path.join(data_path, 'training_data', 'scooby_training_data', 'celltype_fixed.pq'))
cell_type_index['size'] = cell_type_index['cellindex'].apply(lambda x: len(x))
cell_type_index['celltype'] = cell_type_index['celltype'].str.replace(' ', '_').replace(r"G/M_prog", "G+M_prog").replace("MK/E_prog", "MK+E_prog") #+ '_filtered'
cell_type_index = cell_type_index.sort_values('celltype')
cell_type_index.head()

## Load the genes

gtf_file = os.path.join(data_path, "gencode.v32.annotation.sorted.gtf.gz")
transcriptome = Transcriptome(gtf_file)

fasta_file = os.path.join(data_path, 'training_data', "scooby_training_data", "genome_human.fa")
base_path = os.path.join(data_path,'training_data', 'scooby_training_data', 'pseudobulks')
genes_file = os.path.join(data_path,'training_data','scooby_training_data', 'train_val_test_gene_sequences.csv')

gene_coords = pd.read_table(genes_file, header=None)
print(len(gene_coords))
gene_coords.head()

#import anndata
#genes = anndata.read_h5ad('/gstore/data/resbioai/grelu/decima/)

## Get the pseudobulk labels used for training

ds = GenomeIntervalDataset(
    bed_file = genes_file,
    fasta_file = fasta_file,
    filter_df_fn = lambda df: df.filter((pl.col('column_2') >=0)), 
    shift_augs = (0,0),
    return_augs = True,
    context_length = 524288,
)

dataset_targets = onTheFlyPseudobulkDataset(
    cell_types = cell_type_index['celltype'].values,
    ds = ds, 
    base_path = base_path,
)

dataset_target_loader = iter(DataLoader(dataset_targets, batch_size=1, shuffle = False))
print(len(dataset_target_loader))

all_targets = []
gene_names = []

for i in tqdm(range(len(dataset_target_loader))):
    gene_slice, strand = get_gene_slice_and_strand(transcriptome, dataset_targets.genome_ds.df[i, 'column_4'], dataset_targets.genome_ds.df[i, 'column_2'], span = False)
    targets = (next(dataset_target_loader)[2]).float().cuda()
    if len(gene_slice) == 0:
        continue

    if strand == '+':
        t = targets[0, gene_slice, ::2]
    elif strand == '-':
        t = targets[0, gene_slice, 1::2]

    all_targets.append(undo_squashed_scale(t, clip_soft=384).sum(axis=0).detach().clone().cpu().squeeze())
    gene_names.append(val_dataset_targets.genome_ds.df[i, 'column_4'])


all_targets = torch.vstack(all_targets).clone().numpy(force=True)
all_targets.shape

torch.save(all_targets, "count_target_test_no_neighbor.pq")
pd.DataFrame(gene_names).to_parquet("gene_names.pq")
